+++ src/d3d11/d3d11_initializer.h
@@ -2,6 +2,7 @@
 
 #include "d3d11_buffer.h"
 #include "d3d11_texture.h"
+#include "d3d11_view_uav.h"
 
 namespace dxvk {
